refactor: new facade class BotChecker
